Revert "worker concurrency 최적화"
This reverts commit f70cd4c480fc5b57a19ff26c7528350ba690eb90.
