+++ apps/api/src/mq/bullmq.ts
@@ -1,18 +1,14 @@
 import os from 'node:os';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { isMainThread } from 'node:worker_threads';
 import * as Sentry from '@sentry/node';
 import { logger } from '@typie/lib';
 import { Queue, Worker } from 'bullmq';
 import Redis from 'ioredis';
 import { dev, env, stack } from '@/env';
-import { crons } from './tasks';
-
-const __dirname = fileURLToPath(new URL('.', import.meta.url));
+import { crons, jobs } from './tasks';
 
 const log = logger.getChild('mq');
 const lane = dev ? os.hostname() : stack;
+const taskMap = Object.fromEntries([...jobs, ...crons].map((job) => [job.name, job.fn]));
 
 export const queue = new Queue(lane, {
   prefix: `${stack}:{mq}`,
@@ -30,40 +26,31 @@ export const queue = new Queue(lane, {
   },
 });
 
-if (isMainThread) {
-  for (const cron of crons) {
-    queue.upsertJobScheduler(cron.name, {
-      pattern: cron.pattern,
-      tz: 'Asia/Seoul',
-    });
-  }
+if (!process.env.SCRIPT && !process.env.NO_WORKER) {
+  const worker = new Worker(
+    lane,
+    async (job) => {
+      const fn = taskMap[job.name];
+      await fn?.(job.data);
+    },
+    {
+      prefix: `${stack}:{mq}`,
+      connection: new Redis.Cluster([env.REDIS_URL]),
+      concurrency: 50,
+    },
+  );
 
-  if (!process.env.SCRIPT && !process.env.NO_WORKER) {
-    const count = dev ? 1 : 4;
+  worker.on('completed', (job) => {
+    log.info('Job completed {*}', { id: job.id, name: job.name });
+  });
 
-    log.info('Starting workers {*}', { count });
+  worker.on('failed', (job, error) => {
+    log.error('Job failed {*}', { id: job?.id, name: job?.name, error });
+    Sentry.captureException(error);
+  });
 
-    for (let i = 0; i < count; i++) {
-      const worker = new Worker(lane, path.join(__dirname, 'worker.ts'), {
-        prefix: `${stack}:{mq}`,
-        connection: new Redis.Cluster([env.REDIS_URL]),
-        concurrency: 100,
-        useWorkerThreads: true,
-      });
-
-      worker.on('completed', (job) => {
-        log.info('Job completed {*}', { id: job.id, name: job.name });
-      });
-
-      worker.on('failed', (job, error) => {
-        log.error('Job failed {*}', { id: job?.id, name: job?.name, error });
-        Sentry.captureException(error);
-      });
-
-      worker.on('error', (error) => {
-        log.error('Job error {*}', { error });
-        Sentry.captureException(error);
-      });
-    }
-  }
+  worker.on('error', (error) => {
+    log.error('Job error {*}', { error });
+    Sentry.captureException(error);
+  });
 }
