+++ apps/api/src/utils/index.ts
@@ -2,7 +2,6 @@ export * from './auth';
 export * from './date';
 export * from './entity';
 export * from './image-generation';
-export * from './lazy';
 export * from './name';
 export * from './payment';
 export * from './text';
